added: generic|dynamic table
