+++ src/app/features/dashboard/admin/component/product/ProductTable.tsx
@@ -1,13 +1,10 @@
 import { Product, ProductTableProps, Status } from "@/app/types/product";
-import {
-  ColumnDef,
-  getCoreRowModel,
-  useReactTable,
-  flexRender,
-} from "@tanstack/react-table";
+import { ColumnDef } from "@tanstack/react-table";
 import { useMemo } from "react";
+import GenericTable from "../table/GenericTable";
 
 const ProductTable: React.FC<ProductTableProps> = ({ data }) => {
+
   const columns = useMemo<ColumnDef<Product>[]>(
     () => [
       {
@@ -68,50 +65,7 @@ const ProductTable: React.FC<ProductTableProps> = ({ data }) => {
     ],
     [],
   );
-
-  const table = useReactTable({
-    data,
-    columns,
-    getCoreRowModel: getCoreRowModel(),
-  });
-
-  return (
-    <div className="border border-grey-light rounded-lg overflow-x-auto">
-      <table className="min-w-full text-left ">
-        <thead className="bg-base-white">
-          {table.getHeaderGroups().map((headerGroup) => (
-            <tr key={headerGroup.id}>
-              {headerGroup.headers.map((header) => (
-                <th
-                  className="label-bold-14 text-grey px-6 py-4"
-                  key={header.id}
-                >
-                  {flexRender(
-                    header.column.columnDef.header,
-                    header.getContext(),
-                  )}
-                </th>
-              ))}
-            </tr>
-          ))}
-        </thead>
-        <tbody className="divide-y divide-grey-light">
-          {table.getRowModel().rows.map((row) => (
-            <tr key={row.id} className="hover:bg-base-white">
-              {row.getVisibleCells().map((cell) => (
-                <td
-                  className="px-6 py-4   label-regular-14 text-grey-medium"
-                  key={cell.id}
-                >
-                  {flexRender(cell.column.columnDef.cell, cell.getContext())}
-                </td>
-              ))}
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
-  );
+  return <GenericTable columns={columns} data={data} />;
 };
 
 export default ProductTable;
